Reduce duplication in loading schemas
