+++ packages/apollo-cli/src/commands/codegen/generate.ts
@@ -11,10 +11,7 @@ import * as globby from "globby";
 import * as fs from 'fs';
 import { promisify } from 'util';
 
-import { toPromise, execute } from "apollo-link";
-
-import { engineLink, getIdFromKey } from "../../engine";
-import { SCHEMA_QUERY } from "../../operations/schema";
+import { loadSchemaStep } from "../../load-schema";
 
 import { engineFlags } from "../../engine-cli";
 
@@ -32,7 +29,6 @@ export default class Generate extends Command {
     }),
     schema: flags.string({
       description: "Path to your GraphQL schema introspection result",
-      default: "schema.json",
     }),
 
     ...engineFlags,
@@ -120,9 +116,9 @@ export default class Generate extends Command {
     }
 
     const apiKey = flags.key;
-    const pullFromEngine = !!apiKey;
+    const pullFromEngine = !!apiKey && !flags.schema;
 
-    const tasks = new Listr([
+    const tasks: Listr = new Listr([
       {
         title: "Scanning for GraphQL queries",
         task: async (ctx, task) => {
@@ -131,42 +127,19 @@ export default class Generate extends Command {
           ctx.queryPaths = paths;
         }
       },
+      loadSchemaStep(this, pullFromEngine, apiKey, flags.engine, "Loading GraphQL schema", async (ctx) => {
+        const schemaFileContent = await promisify(fs.readFile)(path.resolve(flags.schema as string));
+        const schemaData = JSON.parse(schemaFileContent as any as string);
+        ctx.schema = (schemaData.data) ? schemaData.data.__schema : (
+          schemaData.__schema ? schemaData.__schema : schemaData
+        );
+      }),
       {
-        title: pullFromEngine ? "Loading schema from Apollo Engine" : "Loading GraphQL schema",
-        task: async ctx => {
-          if (pullFromEngine) {
-            const variables = {
-              id: getIdFromKey(apiKey as string),
-              tag: "current",
-            }
-
-            const engineSchema = await toPromise(
-              execute(engineLink, {
-                query: SCHEMA_QUERY,
-                variables,
-                context: {
-                  headers: { ["x-api-key"]: apiKey },
-                  ...(flags.engine && { uri: flags.engine }),
-                },
-              })
-            );
-
-            if (engineSchema.data && engineSchema.data.service.schema) {
-              ctx.schema = buildClientSchema(
-                engineSchema.data.service.schema
-              );
-            } else {
-              this.error("Failed to get schema from Apollo Engine");
-            }
-          } else {
-            const schemaFileContent = await promisify(fs.readFile)(path.resolve(flags.schema as string));
-            const schemaData = JSON.parse(schemaFileContent as any as string);
-            ctx.schema = buildClientSchema(
-              (schemaData.data) ? schemaData.data : (
-                schemaData.__schema ? schemaData : { __schema: schemaData }
-              )
-            )
-          }
+        title: "Parsing GraphQL schema",
+        task: async (ctx) => {
+          ctx.schema = buildClientSchema(
+            { __schema: ctx.schema }
+          );
         }
       },
       {
