+++ packages/apollo-cli/src/commands/schema/checkQueries.ts
@@ -3,22 +3,17 @@ import { Command, flags } from "@oclif/command";
 import * as Listr from "listr";
 import * as path from "path";
 
-import { TargetType, default as generate } from '../../generate';
-
 import { buildClientSchema, validate, findDeprecatedUsages, GraphQLError } from "graphql";
 
 import * as globby from "globby";
 import * as fs from 'fs';
 import { promisify } from 'util';
 
-import { toPromise, execute } from "apollo-link";
-
-import { engineLink, getIdFromKey } from "../../engine";
-import { SCHEMA_QUERY } from "../../operations/schema";
-
 import { loadQueryDocuments } from "apollo-codegen-core/lib/loading";
 
 import { engineFlags } from "../../engine-cli";
+
+import { loadSchemaStep } from "../../load-schema"
 
 // TODO: name this check and rename check to diff?
 export default class CheckQueries extends Command {
@@ -35,7 +30,6 @@ export default class CheckQueries extends Command {
     }),
     schema: flags.string({
       description: "Path to your GraphQL schema introspection result",
-      default: "schema.json",
     }),
 
     ...engineFlags,
@@ -50,9 +44,9 @@ export default class CheckQueries extends Command {
     const { flags } = this.parse(CheckQueries);
 
     const apiKey = flags.key;
-    const pullFromEngine = !!apiKey;
+    const pullFromEngine = !!apiKey && !flags.schema;
 
-    const tasks = new Listr([
+    const tasks: Listr = new Listr([
       {
         title: "Scanning for GraphQL queries",
         task: async (ctx, task) => {
@@ -61,42 +55,19 @@ export default class CheckQueries extends Command {
           ctx.queryPaths = paths;
         }
       },
+      loadSchemaStep(this, pullFromEngine, apiKey, flags.engine, "Loading GraphQL schema", async (ctx) => {
+        const schemaFileContent = await promisify(fs.readFile)(path.resolve(flags.schema as string));
+        const schemaData = JSON.parse(schemaFileContent as any as string);
+        ctx.schema = (schemaData.data) ? schemaData.data.__schema : (
+          schemaData.__schema ? schemaData.__schema : schemaData
+        );
+      }),
       {
-        title: pullFromEngine ? "Loading schema from Apollo Engine" : "Loading GraphQL schema",
-        task: async ctx => {
-          if (pullFromEngine) {
-            const variables = {
-              id: getIdFromKey(apiKey as string),
-              tag: "current",
-            }
-
-            const engineSchema = await toPromise(
-              execute(engineLink, {
-                query: SCHEMA_QUERY,
-                variables,
-                context: {
-                  headers: { ["x-api-key"]: apiKey },
-                  ...(flags.engine && { uri: flags.engine }),
-                },
-              })
-            );
-
-            if (engineSchema.data && engineSchema.data.service.schema) {
-              ctx.schema = buildClientSchema(
-                engineSchema.data.service.schema
-              );
-            } else {
-              this.error("Failed to get schema from Apollo Engine");
-            }
-          } else {
-            const schemaFileContent = await promisify(fs.readFile)(path.resolve(flags.schema as string));
-            const schemaData = JSON.parse(schemaFileContent as any as string);
-            ctx.schema = buildClientSchema(
-              (schemaData.data) ? schemaData.data : (
-                schemaData.__schema ? schemaData : { __schema: schemaData }
-              )
-            )
-          }
+        title: "Parsing GraphQL schema",
+        task: async (ctx) => {
+          ctx.schema = buildClientSchema(
+            { __schema: ctx.schema }
+          );
         }
       },
       {
